requirements.txt
pandas==2.2.2
numpy==1.26.4
pymysql==1.1.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1